Refactor structure
* Break delivery.c into smaller components
