+++ src/CMakeLists.txt
@@ -12,6 +12,16 @@ add_library(stasis_core STATIC
         utils.c
         system.c
         download.c
+        delivery_postprocess.c
+        delivery_conda.c
+        delivery_docker.c
+        delivery_install.c
+        delivery_artifactory.c
+        delivery_test.c
+        delivery_build.c
+        delivery_show.c
+        delivery_populate.c
+        delivery_init.c
         delivery.c
         recipe.c
         relocation.c
